Admin functie toegevoegd

+++ routes/web.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\AboutusController;
 use App\Http\Controllers\EditController;
 use App\Http\Controllers\DeleteController;
 use Illuminate\Support\Facades\Route;
+use App\Http\Middleware\Admin;
 
 /*
 |--------------------------------------------------------------------------
@@ -20,8 +21,8 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/',[PagesController::class, 'home'])->name('home');
 
-Route::get('add-blog-post-form', [PostController::class, 'index'])->middleware('auth');
-Route::post('store-form', [PostController::class, 'store'])->middleware('auth');
+Route::get('add-blog-post-form', [PostController::class, 'index'])->middleware('auth', 'Admin');
+Route::post('store-form', [PostController::class, 'store'])->middleware('auth', 'Admin');
 
 Route::get('contact', [ContactController::class, 'contact'])->name('contact');
 Route::post('contact_form',[ContactController::class, 'contact_form']);
@@ -30,11 +31,9 @@ Route::get('aboutus',[AboutusController::class, 'aboutus'])->name('aboutus');
 
 Route::get('/events', [PagesController::class, 'events'])->name('events');
 
-Route::get('edit/{id}',[EditController::class, 'edit'])->name('edit');
-Route::post('edit/{id}',[EditController::class, 'update'])->name('update');
+Route::get('edit/{id}',[EditController::class, 'edit'])->name('edit')->middleware('auth', 'Admin');
+Route::post('edit/{id}',[EditController::class, 'update'])->name('update')->middleware('auth', 'Admin');
 
-Route::delete('delete/{id}',[DeleteController::class, 'delete'])->name('delete');
-
-
+Route::delete('delete/{id}',[DeleteController::class, 'delete'])->name('delete')->middleware('auth', 'Admin');
 
 require __DIR__.'/auth.php';
